Combine download URL handler inside DownloadHandler to avoid ReferenceError.
